--- feature_audiobook_enhance_details/src/main/java/com/allsoftdroid/audiobook/feature/feature_audiobook_enhance_details/domain/INetworkBaseRepository.kt
@@ -1,6 +1,0 @@
-package com.allsoftdroid.audiobook.feature.feature_audiobook_enhance_details.domain
-
-interface INetworkBaseRepository {
-    fun registerNetworkResponse(listener: NetworkResponseListener)
-    fun unRegisterNetworkResponse()
-}
