+++ feature_audiobook_enhance_details/src/main/java/com/allsoftdroid/audiobook/feature/feature_audiobook_enhance_details/domain/repository/IFetchAdditionBookDetailsRepository.kt
@@ -1,9 +1,10 @@
-package com.allsoftdroid.audiobook.feature.feature_audiobook_enhance_details.domain
+package com.allsoftdroid.audiobook.feature.feature_audiobook_enhance_details.domain.repository
 
 import androidx.lifecycle.LiveData
 import com.allsoftdroid.audiobook.feature.feature_audiobook_enhance_details.data.model.BookDetails
 
-interface IFetchAdditionBookDetailsRepository:INetworkBaseRepository {
+interface IFetchAdditionBookDetailsRepository:
+    INetworkBaseRepository {
     suspend fun fetchBookDetails(bookUrl:String)
     fun getBookDetails(): LiveData<BookDetails>
 }
